Add tests for vector_selector
